+++ app/src/main/java/at/least/conch/MainActivity.kt
@@ -133,7 +133,6 @@ class MainActivity : FragmentActivity() {
                     hostname = p.hostname,
                     port = p.port,
                     username = p.user,
-                    forwardAgent = p.forwardAgent,
                 )
             }
         all.addAll(imported.map { it.second })